# apispec.yml
openapi: "3.0.3"
info:
  title: TechSage Capital API
  version: "1.0"
servers:
  - url: https://api.techsagecapital.nl
    description: Production server
...
paths:
  /auth/login:
    post:
      tags:
        - auth
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                password:
                  type: string
                  description: User's password
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthDto"
        "401":
          description: Invalid credentials supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericMessage"
  /auth/register:
    post:
      tags:
        - auth
      summary: User registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreatePost"
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthDto"
        "400":
          description: Registration failed (e.g., duplicate email)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericMessage"
...
components:
  schemas:
    AuthDto:
      properties:
        jwt:
          type: string
          description: a JSON web token (JWT) required to login.
...
    UserCreatePost:
      type: object
      properties:
        firstName:
          type: string
          description: User's first name
        lastName:
          type: string
          description: User's last name
        email:
          type: string
          format: email
          description: User's email address
        phoneNumber:
          type: string
          description: User's phone number
        bsn:
          type: string
          description: Dutch social security number (BSN)
          example: "000000000"
        password:
          type: string
          description: User's password
...
    GenericMessage:
      type: object
      properties:
        message:
          type: string
          description: Error message
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer